backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0
google-generativeai==0.8.4
python-multipart==0.0.18
yt-dlp==2024.12.23
aiofiles==24.1.0
httpx==0.28.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
